TGP-2890: repackage Normal mode nav files for categorisation
