+++ app/navigation/CategorisationNavigator.scala
@@ -17,12 +17,17 @@
 package navigation
 
 import controllers.routes
+import models.Scenario.getResultAsInt
 import models._
+import models.ott.{CategorisationInfo, CategoryAssessment}
 import pages._
-import pages.categorisation.CategoryGuidancePage
+import pages.categorisation._
 import play.api.mvc.Call
+import queries.{CategorisationDetailsQuery, LongerCategorisationDetailsQuery, LongerCommodityQuery}
 import services.CategorisationService
+import uk.gov.hmrc.play.bootstrap.binders.RedirectUrl
 import utils.Constants.firstAssessmentNumber
+import utils.Constants._
 
 import javax.inject.{Inject, Singleton}
 
@@ -30,14 +35,251 @@ import javax.inject.{Inject, Singleton}
 class CategorisationNavigator @Inject() (categorisationService: CategorisationService) extends NavigatorTrait {
 
   val normalRoutes: Page => UserAnswers => Call = {
+    case p: AssessmentPage                     => navigateFromAssessment(p)
+
     case p: CategoryGuidancePage =>
       _ =>
         controllers.categorisation.routes.AssessmentController.onPageLoad(NormalMode, p.recordId, firstAssessmentNumber)
+    case p: HasSupplementaryUnitPage               => navigateFromHasSupplementaryUnit(p.recordId)
+
+    case p: SupplementaryUnitPage                  =>
+      _ => controllers.categorisation.routes.CyaCategorisationController.onPageLoad(p.recordId)
+    case p: CyaCategorisationPage                  => navigateFromCyaCategorisationPage(p)
+
+    case p: HasCorrectGoodsLongerCommodityCodePage =>
+      answers => navigateFromHasCorrectGoodsLongerCommodityCode(p.recordId, answers, NormalMode)
+    case p: HasSupplementaryUnitUpdatePage         => answers => navigateFromHasSupplementaryUnitUpdatePage(answers, p.recordId)
+    case p: SupplementaryUnitUpdatePage            =>
+      _ => controllers.categorisation.routes.CyaSupplementaryUnitController.onPageLoad(p.recordId)
+    case p: SupplementaryUnitUpdatePage            =>
+      _ => controllers.categorisation.routes.CyaSupplementaryUnitController.onPageLoad(p.recordId)
+    case p: CyaSupplementaryUnitPage               => _ => routes.SingleRecordController.onPageLoad(p.recordId)
+    case p: RecategorisationPreparationPage        => navigateFromReassessmentPrep(p)
+    case p: ReassessmentPage                       => navigateFromReassessment(p)
+    case p: LongerCommodityCodePage                =>
+      _ => routes.HasCorrectGoodsController.onPageLoadLongerCommodityCode(NormalMode, p.recordId)
+    case p: CategorisationPreparationPage          => answers => navigateFromCategorisationPreparationPage(answers, p.recordId)
+
     case _                       => _ => routes.IndexController.onPageLoad()
   }
 
   val checkRoutes: Page => UserAnswers => Call = { case _ =>
     _ => routes.JourneyRecoveryController.onPageLoad()
   }
+  private def navigateFromCategorisationPreparationPage(answers: UserAnswers, recordId: String): Call =
+    answers.get(CategorisationDetailsQuery(recordId)) match {
+      case Some(catInfo) if catInfo.isCommCodeExpired                           =>
+        routes.ExpiredCommodityCodeController.onPageLoad(recordId)
+      case Some(catInfo) if catInfo.categoryAssessmentsThatNeedAnswers.nonEmpty =>
+        controllers.categorisation.routes.CategoryGuidanceController.onPageLoad(recordId)
+
+      case Some(catInfo) =>
+        val scenario = categorisationService.calculateResult(catInfo, answers, recordId)
+
+        if (shouldGoToLongerCommodityCodeFromPrepPage(catInfo, scenario)) {
+          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
+        } else if (shouldGoToSupplementaryUnitFromPrepPage(catInfo, scenario)) {
+          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
+        } else {
+          controllers.categorisation.routes.CategorisationResultController.onPageLoad(recordId, scenario)
+        }
+
+      case None => routes.JourneyRecoveryController.onPageLoad()
+    }
+
+  private def getShortenedCommodityCode(commodity: Commodity) =
+    commodity.commodityCode.reverse
+      .dropWhile(char => char == '0')
+      .reverse
+      .padTo(minimumLengthOfCommodityCode, '0')
+      .mkString
+
+  private def shouldGoToLongerCommodityCodeFromPrepPage(catInfo: CategorisationInfo, scenario: Scenario) =
+    catInfo.getMinimalCommodityCode.length == minimumLengthOfCommodityCode && catInfo.descendantCount != 0 &&
+      getResultAsInt(scenario) == Category2AsInt
+
+  private def navigateFromHasCorrectGoodsLongerCommodityCode(
+                                                              recordId: String,
+                                                              answers: UserAnswers,
+                                                              mode: Mode
+                                                            ): Call = {
+    for {
+      categorisationInfo <- answers.get(CategorisationDetailsQuery(recordId))
+      commodity          <- answers.get(LongerCommodityQuery(recordId))
+    } yield
+      if (categorisationInfo.commodityCode == getShortenedCommodityCode(commodity)) {
+        controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(mode, recordId)
+      } else {
+        answers.get(HasCorrectGoodsLongerCommodityCodePage(recordId)) match {
+          case Some(true)  =>
+            controllers.categorisation.routes.CategorisationPreparationController
+              .startLongerCategorisation(mode, recordId)
+          case Some(false) => controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(mode, recordId)
+          case None        => routes.JourneyRecoveryController.onPageLoad()
+        }
+      }
+  }.getOrElse(routes.JourneyRecoveryController.onPageLoad())
+  private def navigateFromReassessment(assessmentPage: ReassessmentPage)(answers: UserAnswers): Call = {
+    val recordId   = assessmentPage.recordId
+    val nextIndex  = assessmentPage.index + 1
+    val nextNumber = nextIndex + 1
+
+    {
+      for {
+        categorisationInfo <- answers.get(LongerCategorisationDetailsQuery(recordId))
+        assessmentCount     = categorisationInfo.categoryAssessmentsThatNeedAnswers.size
+        assessmentQuestion <- categorisationInfo.getAssessmentFromIndex(assessmentPage.index)
+        nextAnswer          = answers.get(ReassessmentPage(recordId, nextIndex))
+        assessmentAnswer   <- answers.get(assessmentPage)
+      } yield assessmentAnswer.answer match {
+        case AssessmentAnswer.Exemption(_)
+          if nextIndex < assessmentCount && (reassessmentAnswerIsEmpty(nextAnswer) || !nextAnswer.exists(
+            _.isAnswerCopiedFromPreviousAssessment
+          )) =>
+          controllers.categorisation.routes.AssessmentController
+            .onPageLoadReassessment(NormalMode, recordId, nextNumber)
+        case AssessmentAnswer.Exemption(_) if nextIndex < assessmentCount                                        =>
+          navigateFromReassessment(ReassessmentPage(recordId, nextIndex))(answers)
+        case AssessmentAnswer.NoExemption if shouldGoToSupplementaryUnit(categorisationInfo, assessmentQuestion) =>
+          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
+        case _                                                                                                   =>
+          controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
+      }
+    }.getOrElse(routes.JourneyRecoveryController.onPageLoad())
+  }
+
+  private def reassessmentAnswerIsEmpty(nextAnswer: Option[ReassessmentAnswer]) =
+    nextAnswer.isEmpty || nextAnswer.exists(reassessment => reassessment.answer == AssessmentAnswer.NotAnsweredYet)
+
+  private def navigateFromReassessmentPrep(
+                                            reasessmentPrep: RecategorisationPreparationPage
+                                          )(answers: UserAnswers): Call = {
+    val recordId = reasessmentPrep.recordId
+
+    answers.get(LongerCategorisationDetailsQuery(recordId)) match {
+      case Some(catInfo) if catInfo.categoryAssessmentsThatNeedAnswers.nonEmpty =>
+        val firstAnswer = answers.get(ReassessmentPage(recordId, firstAssessmentIndex))
+        if (reassessmentAnswerIsEmpty(firstAnswer) || !firstAnswer.get.isAnswerCopiedFromPreviousAssessment) {
+          controllers.categorisation.routes.AssessmentController
+            .onPageLoadReassessment(NormalMode, recordId, firstAssessmentNumber)
+        } else {
+          navigateFromReassessment(ReassessmentPage(recordId, firstAssessmentIndex))(answers)
+        }
+
+      case Some(catInfo) =>
+        val scenario = categorisationService.calculateResult(catInfo, answers, recordId)
+        if (shouldGoToSupplementaryUnitFromPrepPage(catInfo, scenario)) {
+          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
+        } else {
+          controllers.categorisation.routes.CategorisationResultController.onPageLoad(recordId, scenario)
+        }
+
+      case None => routes.JourneyRecoveryController.onPageLoad()
+
+    }
+  }
+
+  private def navigateFromHasSupplementaryUnitUpdatePage(answers: UserAnswers, recordId: String): Call = {
+    val continueUrl = RedirectUrl(routes.SingleRecordController.onPageLoad(recordId).url)
+    answers
+      .get(HasSupplementaryUnitUpdatePage(recordId))
+      .map {
+        case false => controllers.categorisation.routes.CyaSupplementaryUnitController.onPageLoad(recordId)
+        case true  =>
+          controllers.categorisation.routes.SupplementaryUnitController.onPageLoadUpdate(NormalMode, recordId)
+      }
+      .getOrElse(routes.JourneyRecoveryController.onPageLoad(Some(continueUrl)))
+  }
+
+
+  private def navigateFromCyaCategorisationPage(page: CyaCategorisationPage)(answers: UserAnswers): Call =
+    (for {
+      categorisationInfo <- getCategorisationInfoFromUA(answers, page.recordId)
+      scenario            = categorisationService.calculateResult(categorisationInfo, answers, page.recordId)
+    } yield controllers.categorisation.routes.CategorisationResultController.onPageLoad(page.recordId, scenario))
+      .getOrElse(
+        routes.JourneyRecoveryController.onPageLoad(
+          Some(
+            RedirectUrl(
+              controllers.categorisation.routes.CategorisationPreparationController
+                .startCategorisation(page.recordId)
+                .url
+            )
+          )
+        )
+      )
+
+  private def getCategorisationInfoFromUA(userAnswers: UserAnswers, recordId: String) = {
+    val longerDetails = userAnswers.get(LongerCategorisationDetailsQuery(recordId))
+    longerDetails match {
+      case Some(_) => longerDetails
+      case _       => userAnswers.get(CategorisationDetailsQuery(recordId))
+    }
+  }
+
+
+
+
+
+  private def navigateFromHasSupplementaryUnit(recordId: String)(answers: UserAnswers): Call =
+    answers
+      .get(HasSupplementaryUnitPage(recordId))
+      .map {
+        case true  => controllers.categorisation.routes.SupplementaryUnitController.onPageLoad(NormalMode, recordId)
+        case false => controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
+      }
+      .getOrElse(routes.JourneyRecoveryController.onPageLoad())
+
+
+  private def shouldGoToSupplementaryUnit(
+                                           categorisationInfo: CategorisationInfo,
+                                           assessmentQuestion: CategoryAssessment
+                                         ) =
+    assessmentQuestion.category == Category2AsInt && categorisationInfo.measurementUnit.isDefined
+
+  private def shouldGoToSupplementaryUnitFromPrepPage(catInfo: CategorisationInfo, scenario: Scenario) =
+    catInfo.measurementUnit.isDefined && getResultAsInt(scenario) == Category2AsInt
+
+  private def shouldGoToLongerCommodityCodeWhenCategory2(
+                                                          categorisationInfo: CategorisationInfo,
+                                                          assessmentQuestion: CategoryAssessment
+                                                        ) =
+    categorisationInfo.getMinimalCommodityCode.length == 6 && categorisationInfo.descendantCount != 0 && assessmentQuestion.category == Category2AsInt
+
+  private def shouldGoToLongerCommodityCodeWhenCategory1(
+                                                          categorisationInfo: CategorisationInfo,
+                                                          assessmentQuestion: CategoryAssessment
+                                                        ) =
+    categorisationInfo.getMinimalCommodityCode.length == minimumLengthOfCommodityCode && categorisationInfo.descendantCount != 0 && assessmentQuestion.category == Category1AsInt && !categorisationInfo.categoryAssessmentsThatNeedAnswers
+      .exists(_.isCategory2) && categorisationInfo.categoryAssessments.exists(_.isCategory2)
+
+  private def navigateFromAssessment(assessmentPage: AssessmentPage)(answers: UserAnswers): Call = {
+    val recordId   = assessmentPage.recordId
+    val nextIndex  = assessmentPage.index + 1
+    val nextNumber = nextIndex + 1
+
+    {
+      for {
+        categorisationInfo <- answers.get(CategorisationDetailsQuery(recordId))
+        assessmentCount     = categorisationInfo.categoryAssessmentsThatNeedAnswers.size
+        assessmentQuestion <- categorisationInfo.getAssessmentFromIndex(assessmentPage.index)
+        assessmentAnswer   <- answers.get(assessmentPage)
+      } yield assessmentAnswer match {
+        case AssessmentAnswer.Exemption(_) if nextIndex < assessmentCount                                        =>
+          controllers.categorisation.routes.AssessmentController.onPageLoad(NormalMode, recordId, nextNumber)
+        case AssessmentAnswer.Exemption(_)
+          if shouldGoToLongerCommodityCodeWhenCategory1(categorisationInfo, assessmentQuestion) =>
+          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
+        case AssessmentAnswer.NoExemption
+          if shouldGoToLongerCommodityCodeWhenCategory2(categorisationInfo, assessmentQuestion) =>
+          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
+        case AssessmentAnswer.NoExemption if shouldGoToSupplementaryUnit(categorisationInfo, assessmentQuestion) =>
+          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
+        case _                                                                                                   =>
+          controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
+      }
+    }.getOrElse(routes.JourneyRecoveryController.onPageLoad())
+  }
+
 
 }
