+++ app/navigation/Navigator.scala
@@ -22,7 +22,7 @@ import models.Scenario.getResultAsInt
 import models._
 import models.ott.{CategorisationInfo, CategoryAssessment}
 import pages._
-import pages.categorisation.{AssessmentPage, CategorisationPreparationPage, CategoryGuidancePage, CyaCategorisationPage, CyaSupplementaryUnitPage, HasSupplementaryUnitPage, HasSupplementaryUnitUpdatePage, ReassessmentPage, SupplementaryUnitPage}
+import pages.categorisation._
 import play.api.mvc.Call
 import queries.{CategorisationDetailsQuery, LongerCategorisationDetailsQuery, LongerCommodityQuery}
 import services.CategorisationService
@@ -46,38 +46,19 @@ class Navigator @Inject() (categorisationService: CategorisationService) extends
     case p: CommodityCodeUpdatePage                => _ => routes.HasCorrectGoodsController.onPageLoadUpdate(NormalMode, p.recordId)
     case HasCorrectGoodsPage                       => answers => navigateFromHasCorrectGoods(answers)
     case p: HasCorrectGoodsCommodityCodeUpdatePage => answers => navigateFromHasCorrectGoodsUpdate(answers, p.recordId)
-    case p: AssessmentPage                         => navigateFromAssessment(p)
-    case p: HasSupplementaryUnitPage               => navigateFromHasSupplementaryUnit(p.recordId)
-    case p: SupplementaryUnitPage                  =>
-      _ => controllers.categorisation.routes.CyaCategorisationController.onPageLoad(p.recordId)
     case p: AdviceStartPage                        => _ => routes.NameController.onPageLoad(NormalMode, p.recordId)
     case p: NamePage                               => _ => routes.EmailController.onPageLoad(NormalMode, p.recordId)
     case p: EmailPage                              => _ => routes.CyaRequestAdviceController.onPageLoad(p.recordId)
-    case p: CyaCategorisationPage                  => navigateFromCyaCategorisationPage(p)
     case RemoveGoodsRecordPage                     => _ => routes.GoodsRecordsController.onPageLoad(firstPage)
-    case p: LongerCommodityCodePage                =>
-      _ => routes.HasCorrectGoodsController.onPageLoadLongerCommodityCode(NormalMode, p.recordId)
-    case p: HasCorrectGoodsLongerCommodityCodePage =>
-      answers => navigateFromHasCorrectGoodsLongerCommodityCode(p.recordId, answers, NormalMode)
     case p: HasGoodsDescriptionChangePage          => answers => navigateFromHasGoodsDescriptionChangePage(answers, p.recordId)
     case p: HasCountryOfOriginChangePage           => answers => navigateFromHasCountryOfOriginChangePage(answers, p.recordId)
     case p: HasCommodityCodeChangePage             => answers => navigateFromHasCommodityCodeChangePage(answers, p.recordId)
-    case p: HasSupplementaryUnitUpdatePage         => answers => navigateFromHasSupplementaryUnitUpdatePage(answers, p.recordId)
-    case p: SupplementaryUnitUpdatePage            =>
-      _ => controllers.categorisation.routes.CyaSupplementaryUnitController.onPageLoad(p.recordId)
     case p: ReviewReasonPage                       => _ => routes.SingleRecordController.onPageLoad(p.recordId)
-    case p: CategorisationPreparationPage          => answers => navigateFromCategorisationPreparationPage(answers, p.recordId)
-    case p: CategoryGuidancePage                   =>
-      _ =>
-        controllers.categorisation.routes.AssessmentController.onPageLoad(NormalMode, p.recordId, firstAssessmentNumber)
-    case p: ReassessmentPage                       => navigateFromReassessment(p)
-    case p: RecategorisationPreparationPage        => navigateFromReassessmentPrep(p)
     case p: WithdrawAdviceStartPage                => answers => navigateFromWithdrawAdviceStartPage(answers, p.recordId)
     case p: ReasonForWithdrawAdvicePage            => _ => routes.WithdrawAdviceSuccessController.onPageLoad(p.recordId)
     case p: CyaCreateRecordPage                    => _ => routes.CreateRecordSuccessController.onPageLoad(p.recordId)
     case p: CyaRequestAdvicePage                   => _ => routes.AdviceSuccessController.onPageLoad(p.recordId)
     case p: CyaUpdateRecordPage                    => _ => routes.SingleRecordController.onPageLoad(p.recordId)
-    case p: CyaSupplementaryUnitPage               => _ => routes.SingleRecordController.onPageLoad(p.recordId)
     case PreviousMovementRecordsPage               => _ => routes.GoodsRecordsController.onPageLoad(firstPage)
     case RequestDataPage                           => _ => routes.DownloadRequestSuccessController.onPageLoad()
     case _                                         => _ => routes.IndexController.onPageLoad()
@@ -148,18 +129,6 @@ class Navigator @Inject() (categorisationService: CategorisationService) extends
       .getOrElse(routes.JourneyRecoveryController.onPageLoad(Some(continueUrl)))
   }
 
-  private def navigateFromHasSupplementaryUnitUpdatePage(answers: UserAnswers, recordId: String): Call = {
-    val continueUrl = RedirectUrl(routes.SingleRecordController.onPageLoad(recordId).url)
-    answers
-      .get(HasSupplementaryUnitUpdatePage(recordId))
-      .map {
-        case false => controllers.categorisation.routes.CyaSupplementaryUnitController.onPageLoad(recordId)
-        case true  =>
-          controllers.categorisation.routes.SupplementaryUnitController.onPageLoadUpdate(NormalMode, recordId)
-      }
-      .getOrElse(routes.JourneyRecoveryController.onPageLoad(Some(continueUrl)))
-  }
-
   private def navigateFromHasGoodsDescriptionChangePage(answers: UserAnswers, recordId: String): Call = {
     val continueUrl = RedirectUrl(routes.SingleRecordController.onPageLoad(recordId).url)
     answers
@@ -218,127 +187,8 @@ class Navigator @Inject() (categorisationService: CategorisationService) extends
       }
       .getOrElse(routes.JourneyRecoveryController.onPageLoad())
 
-  private def navigateFromHasSupplementaryUnit(recordId: String)(answers: UserAnswers): Call =
-    answers
-      .get(HasSupplementaryUnitPage(recordId))
-      .map {
-        case true  => controllers.categorisation.routes.SupplementaryUnitController.onPageLoad(NormalMode, recordId)
-        case false => controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
-      }
-      .getOrElse(routes.JourneyRecoveryController.onPageLoad())
-
-  private def navigateFromCategorisationPreparationPage(answers: UserAnswers, recordId: String): Call =
-    answers.get(CategorisationDetailsQuery(recordId)) match {
-      case Some(catInfo) if catInfo.isCommCodeExpired                           =>
-        routes.ExpiredCommodityCodeController.onPageLoad(recordId)
-      case Some(catInfo) if catInfo.categoryAssessmentsThatNeedAnswers.nonEmpty =>
-        controllers.categorisation.routes.CategoryGuidanceController.onPageLoad(recordId)
-
-      case Some(catInfo) =>
-        val scenario = categorisationService.calculateResult(catInfo, answers, recordId)
-
-        if (shouldGoToLongerCommodityCodeFromPrepPage(catInfo, scenario)) {
-          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
-        } else if (shouldGoToSupplementaryUnitFromPrepPage(catInfo, scenario)) {
-          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
-        } else {
-          controllers.categorisation.routes.CategorisationResultController.onPageLoad(recordId, scenario)
-        }
-
-      case None => routes.JourneyRecoveryController.onPageLoad()
-    }
-
-  private def shouldGoToLongerCommodityCodeFromPrepPage(catInfo: CategorisationInfo, scenario: Scenario) =
-    catInfo.getMinimalCommodityCode.length == minimumLengthOfCommodityCode && catInfo.descendantCount != 0 &&
-      getResultAsInt(scenario) == Category2AsInt
-
   private def shouldGoToSupplementaryUnitFromPrepPage(catInfo: CategorisationInfo, scenario: Scenario) =
     catInfo.measurementUnit.isDefined && getResultAsInt(scenario) == Category2AsInt
-
-  private def navigateFromReassessmentPrep(
-    reasessmentPrep: RecategorisationPreparationPage
-  )(answers: UserAnswers): Call = {
-    val recordId = reasessmentPrep.recordId
-
-    answers.get(LongerCategorisationDetailsQuery(recordId)) match {
-      case Some(catInfo) if catInfo.categoryAssessmentsThatNeedAnswers.nonEmpty =>
-        val firstAnswer = answers.get(ReassessmentPage(recordId, firstAssessmentIndex))
-        if (reassessmentAnswerIsEmpty(firstAnswer) || !firstAnswer.get.isAnswerCopiedFromPreviousAssessment) {
-          controllers.categorisation.routes.AssessmentController
-            .onPageLoadReassessment(NormalMode, recordId, firstAssessmentNumber)
-        } else {
-          navigateFromReassessment(ReassessmentPage(recordId, firstAssessmentIndex))(answers)
-        }
-
-      case Some(catInfo) =>
-        val scenario = categorisationService.calculateResult(catInfo, answers, recordId)
-        if (shouldGoToSupplementaryUnitFromPrepPage(catInfo, scenario)) {
-          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
-        } else {
-          controllers.categorisation.routes.CategorisationResultController.onPageLoad(recordId, scenario)
-        }
-
-      case None => routes.JourneyRecoveryController.onPageLoad()
-
-    }
-  }
-
-  private def navigateFromReassessment(assessmentPage: ReassessmentPage)(answers: UserAnswers): Call = {
-    val recordId   = assessmentPage.recordId
-    val nextIndex  = assessmentPage.index + 1
-    val nextNumber = nextIndex + 1
-
-    {
-      for {
-        categorisationInfo <- answers.get(LongerCategorisationDetailsQuery(recordId))
-        assessmentCount     = categorisationInfo.categoryAssessmentsThatNeedAnswers.size
-        assessmentQuestion <- categorisationInfo.getAssessmentFromIndex(assessmentPage.index)
-        nextAnswer          = answers.get(ReassessmentPage(recordId, nextIndex))
-        assessmentAnswer   <- answers.get(assessmentPage)
-      } yield assessmentAnswer.answer match {
-        case AssessmentAnswer.Exemption(_)
-            if nextIndex < assessmentCount && (reassessmentAnswerIsEmpty(nextAnswer) || !nextAnswer.exists(
-              _.isAnswerCopiedFromPreviousAssessment
-            )) =>
-          controllers.categorisation.routes.AssessmentController
-            .onPageLoadReassessment(NormalMode, recordId, nextNumber)
-        case AssessmentAnswer.Exemption(_) if nextIndex < assessmentCount                                        =>
-          navigateFromReassessment(ReassessmentPage(recordId, nextIndex))(answers)
-        case AssessmentAnswer.NoExemption if shouldGoToSupplementaryUnit(categorisationInfo, assessmentQuestion) =>
-          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
-        case _                                                                                                   =>
-          controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
-      }
-    }.getOrElse(routes.JourneyRecoveryController.onPageLoad())
-  }
-
-  private def navigateFromAssessment(assessmentPage: AssessmentPage)(answers: UserAnswers): Call = {
-    val recordId   = assessmentPage.recordId
-    val nextIndex  = assessmentPage.index + 1
-    val nextNumber = nextIndex + 1
-
-    {
-      for {
-        categorisationInfo <- answers.get(CategorisationDetailsQuery(recordId))
-        assessmentCount     = categorisationInfo.categoryAssessmentsThatNeedAnswers.size
-        assessmentQuestion <- categorisationInfo.getAssessmentFromIndex(assessmentPage.index)
-        assessmentAnswer   <- answers.get(assessmentPage)
-      } yield assessmentAnswer match {
-        case AssessmentAnswer.Exemption(_) if nextIndex < assessmentCount                                        =>
-          controllers.categorisation.routes.AssessmentController.onPageLoad(NormalMode, recordId, nextNumber)
-        case AssessmentAnswer.Exemption(_)
-            if shouldGoToLongerCommodityCodeWhenCategory1(categorisationInfo, assessmentQuestion) =>
-          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
-        case AssessmentAnswer.NoExemption
-            if shouldGoToLongerCommodityCodeWhenCategory2(categorisationInfo, assessmentQuestion) =>
-          controllers.categorisation.routes.LongerCommodityCodeController.onPageLoad(NormalMode, recordId)
-        case AssessmentAnswer.NoExemption if shouldGoToSupplementaryUnit(categorisationInfo, assessmentQuestion) =>
-          controllers.categorisation.routes.HasSupplementaryUnitController.onPageLoad(NormalMode, recordId)
-        case _                                                                                                   =>
-          controllers.categorisation.routes.CyaCategorisationController.onPageLoad(recordId)
-      }
-    }.getOrElse(routes.JourneyRecoveryController.onPageLoad())
-  }
 
   private def shouldGoToLongerCommodityCodeWhenCategory2(
     categorisationInfo: CategorisationInfo,
@@ -352,37 +202,6 @@ class Navigator @Inject() (categorisationService: CategorisationService) extends
   ) =
     categorisationInfo.getMinimalCommodityCode.length == minimumLengthOfCommodityCode && categorisationInfo.descendantCount != 0 && assessmentQuestion.category == Category1AsInt && !categorisationInfo.categoryAssessmentsThatNeedAnswers
       .exists(_.isCategory2) && categorisationInfo.categoryAssessments.exists(_.isCategory2)
-
-  private def shouldGoToSupplementaryUnit(
-    categorisationInfo: CategorisationInfo,
-    assessmentQuestion: CategoryAssessment
-  ) =
-    assessmentQuestion.category == Category2AsInt && categorisationInfo.measurementUnit.isDefined
-
-  private def navigateFromCyaCategorisationPage(page: CyaCategorisationPage)(answers: UserAnswers): Call =
-    (for {
-      categorisationInfo <- getCategorisationInfoFromUA(answers, page.recordId)
-      scenario            = categorisationService.calculateResult(categorisationInfo, answers, page.recordId)
-    } yield controllers.categorisation.routes.CategorisationResultController.onPageLoad(page.recordId, scenario))
-      .getOrElse(
-        routes.JourneyRecoveryController.onPageLoad(
-          Some(
-            RedirectUrl(
-              controllers.categorisation.routes.CategorisationPreparationController
-                .startCategorisation(page.recordId)
-                .url
-            )
-          )
-        )
-      )
-
-  private def getCategorisationInfoFromUA(userAnswers: UserAnswers, recordId: String) = {
-    val longerDetails = userAnswers.get(LongerCategorisationDetailsQuery(recordId))
-    longerDetails match {
-      case Some(_) => longerDetails
-      case _       => userAnswers.get(CategorisationDetailsQuery(recordId))
-    }
-  }
 
   private def navigateFromHasCorrectGoodsCheck(answers: UserAnswers): Call =
     answers
